wallet_manager/requirements.txt
arrow==0.12.1
async-timeout==3.0.0
atomicwrites==1.1.5
attrdict==2.0.0
attrs==18.1.0
base58==0.2.2
base58check==1.0.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
cytoolz==0.9.0.1
docker-py==1.8.0
ecdsa==0.11
eth-abi==1.1.1
eth-account==0.3.0
eth-hash==0.1.4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-utils==1.0.3
flake8==3.5.0
funcsigs==1.0.2
future==0.16.0
hexbytes==0.1.0
idna==2.7
idna-ssl==1.1.0
jsonrpcclient==2.6.0
jsonrpcserver==3.5.6
jsonschema==2.6.0
lru-dict==1.1.6
mccabe==0.6.1
mnemonic==0.13
more-itertools==4.2.0
multidict==4.3.1
parsimonious==0.8.0
path.py==11.0.1
pbkdf2==1.3
pexpect==4.6.0
pkg-resources==0.0.0
pluggy==0.6.0
protobuf==3.0.0a3
ptyprocess==0.6.0
py==1.5.4
py-solc==3.1.0
pyaes==1.6.1
pycodestyle==2.3.1
pycryptodome==3.6.4
pyflakes==1.6.0
pysha3==1.0.2
pytest==3.6.3
python-bitcoinrpc==1.0
python-dateutil==2.7.3
pywallet==0.0.7
PyYAML==3.13
requests==2.19.1
rlp==0.6.0
semantic-version==2.6.0
sha256==0.1
six==1.11.0
tabulate==0.8.2
toolz==0.9.0
tornado==5.1
two1==3.10.9
urllib3==1.23
web3==4.5.0
websocket-client==0.48.0
websockets==5.0.1
yarl==1.2.6
